+++ Source/NSObject.m
@@ -1840,7 +1840,8 @@ static id gs_weak_load(id obj)
         [NSException
 	  raise: NSGenericException
 	  format: @"Autorelease would release object too many times.\n"
-	  @"%d release(s) versus %d retain(s)", release_count, retain_count];
+	  @"%lu release(s) versus %lu retain(s)",
+	  (unsigned long)release_count, (unsigned long)retain_count];
     }
 
   (*autorelease_imp)(autorelease_class, autorelease_sel, self);
